feat : finished webpage + added button on main webpage
