--- ym-jf_energy-mix/README.md
@@ -1,11 +1,0 @@
-# Energy mix of different countries
-
-source : https://www.kaggle.com/datasets/pralabhpoudel/world-energy-consumption
-
-The dataset used describes the energy mix of every country in the world from 1900 to 2019 and 
-the variations for every energy production type.
-
-things to do ideas
-* Energy mix over a century graph for every country
-* Heatmap of nuclear, coal energy mix share
-* Evolution of coal use in different continents
